Key-Value Pair + Safe Approach on how data is retrieved
Member functions no longer return pointers to nodes because someone could compromise the balancing or even the validity of the binary search tree data structure by modifying the returned node. So what i did is i made sure the end user can check the return value to see if the node was found and then only if it was found, the programmer proceeds using the returned key or value. If i could come up with a better solution i would. Any suggestions are welcome!